Implement --fiat to support fiat files by argument
`--fiat` wasn't actually implemented despite being reported so by the
help.

I implemented this by modifying the `FIAT_FILE` env var of the running
script if --fiat is used. After changing the env var, the config module
is hot reloaded to pick up the latest `FIAT_FILE`.

Next, I updated `albumClass` and `playlistClass` to accept a fiat file
by argument. `Decouple`'s config is immutable and `playlistClass` is
imported early on in the main script. Thus, album and playlist don't see
the latest `FIAT_FILE` changes. Passing it by argument is cleaner anyway
since that reduces the strong coupling between the modules. Plus,
without that new argument then the config would need to be reloaded on
each new instance of `playlistClass` - this is messy and wasteful.

The `fiat_file` parameter defaults to whatever is in config so the
classes weren't changed all that much.

A better solution would need a deeper rewrite. And I don't want to break
the script. I don't want to be known as that dummy that broke Rivers
Cuomo's code.